refactor(folderStructure): tests in module subfolder
