+++ packages/core/manifest/src/app.module.ts
@@ -9,7 +9,7 @@ import generalConfig from './config/general'
 import pathsConfig from './config/paths'
 import { CrudModule } from './crud/crud.module'
 import { EntityModule } from './entity/entity.module'
-import { EntityLoaderService } from './entity/services/entity-loader/entity-loader.service'
+import { EntityLoaderService } from './entity/services/entity-loader.service'
 import { LoggerModule } from './logger/logger.module'
 import { LoggerService } from './logger/logger.service'
 import { ManifestModule } from './manifest/manifest.module'
