+++ packages/core/manifest/src/entity/services/relationship.service.ts
@@ -1,7 +1,7 @@
 import { faker } from '@faker-js/faker'
 import { EntityManifest, RelationshipManifest } from '@mnfst/types'
 import { Injectable } from '@nestjs/common'
-import { ManifestService } from '../../../manifest/services/manifest/manifest.service'
+import { ManifestService } from '../../manifest/services/manifest.service'
 
 @Injectable()
 export class RelationshipService {
