+++ packages/core/manifest/src/manifest/constants.ts
