+++ packages/core/manifest/src/manifest/manifest.module.ts
@@ -2,9 +2,9 @@ import { Module, forwardRef } from '@nestjs/common'
 import { AuthService } from '../auth/auth.service'
 import { EntityModule } from '../entity/entity.module'
 import { ManifestController } from './controllers/manifest.controller'
-import { ManifestService } from './services/manifest/manifest.service'
-import { SchemaService } from './services/schema/schema.service'
-import { YamlService } from './services/yaml/yaml.service'
+import { ManifestService } from './services/manifest.service'
+import { SchemaService } from './services/schema.service'
+import { YamlService } from './services/yaml.service'
 
 /**
  *
