+++ packages/core/manifest/src/manifest/services/yaml.service.ts
